rename package pkg/auth/signing to pkg/signing
