+++ pkg/signing/encode.go
