+++ pkg/signing/peers.go
