+++ pkg/signing/verify.go
