+++ storagenode/storagenodedb/storagenodedbtest/run_test.go
@@ -17,8 +17,8 @@ import (
 	"storj.io/storj/internal/testidentity"
 	"storj.io/storj/internal/testrand"
 	"storj.io/storj/internal/teststorj"
-	"storj.io/storj/pkg/auth/signing"
 	"storj.io/storj/pkg/pb"
+	"storj.io/storj/pkg/signing"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/storagenode"
 	"storj.io/storj/storagenode/orders"
