+++ storagenode/vouchers/verification.go
@@ -10,8 +10,8 @@ import (
 	"github.com/zeebo/errs"
 
 	"storj.io/storj/internal/errs2"
-	"storj.io/storj/pkg/auth/signing"
 	"storj.io/storj/pkg/pb"
+	"storj.io/storj/pkg/signing"
 	"storj.io/storj/pkg/storj"
 )
